pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "event-driven-video-metadata-platform"
version = "0.1.0"
description = "Event-driven video metadata pipeline (FastAPI + AWS + Terraform)"
requires-python = ">=3.11"
dependencies = [
  "boto3>=1.34.0",
  "fastapi>=0.115.0",
  "httpx>=0.27.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "prometheus-client>=0.20.0",
  "python-json-logger>=2.0.7",
  "PyJWT>=2.9.0",
  "redis>=5.0.8",
  "tenacity>=9.0.0",
  "uvicorn[standard]>=0.30.0",
]

[project.optional-dependencies]
dev = [
  "bandit>=1.7.9",
  "mypy>=1.11.0",
  "pip-audit>=2.7.3",
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.8",
  "pytest-cov>=5.0.0",
  "respx>=0.21.1",
  "ruff>=0.6.4",
  "types-redis>=4.6.0.20241004",
  "types-PyYAML>=6.0.12.20240917",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
no_implicit_optional = true
ignore_missing_imports = true
